- Add custom User class model inherited from IdentityUser - Scaffold again the Net Core Identity pages to use the custom user class. - Add many-to-many relationship between items and categories. - Add new migrations (V3 latest) to reflect the new database changes.

+++ Data/Models/Order.cs
@@ -18,7 +18,7 @@ namespace FarmaShop.Data.Models
 
         public string Country { get; set; }
 
-        public decimal OrderTotal { get; set; }
+        public double OrderTotal { get; set; }
 
         public DateTime OrderPlaced { get; set; }
